refactor: remove MCTS GPU backend
