+++ package.json
@@ -25,9 +25,6 @@
   "dependencies": {
     "http-server": "^14.1.1"
   },
-  "optionalDependencies": {
-    "gpu.js": "^2.15.0"
-  },
   "devDependencies": {
     "@eslint/js": "^9.35.0",
     "sharp": "^0.33.4",
